Treat numbers with zero fraction as valid integers
